win_requirements.txt
colorama


pillow

PyInstaller


pycryptodome
pyreadline
pywin32
requests